feat: network type recognition for each node
Allow recognize type of network for each node. Currently it allows to check if
node is from real ZCash network or from any other.

Classification is based on three factors:
start_height - this MUST match
port
agent string

If either port or agent string would not match node still can be considered as
ZCash node as long as start_height matches. The only exception is when agent string
identifies node as version 6 (which is other network than ZCash) - that is blocker and
even when two other conditions are met, node wouldn't be considered as valid ZCash node.