feat: add default_port argument to crawler command line
Add new command line argument which allows user to specify default port
to be used with seeds with no port specified. It allows running crawler
both with mainnet and testnet as well as nodes running on non-standard
ports.

Note: there is still a need to change network MAGIC value to switch
between mainnet and testnet.